% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-run.R
\name{app-run}
\alias{app-run}
\alias{runPaletteApp}
\alias{app_server}
\alias{app_ui}
\title{Run App}
\usage{
runPaletteApp()

app_server()

app_ui()
}
\description{
Functions to run the package shiny app
}
\section{Functions}{
\itemize{
\item \code{runPaletteApp()}: returns app object for subsequent execution

\item \code{app_server()}: server function for app

\item \code{app_ui()}: UI function for app

}}
