% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{app-utils}
\alias{app-utils}
\alias{new_waiter}
\alias{waiter_html}
\alias{is_ci}
\alias{is_testing}
\alias{get_app_colors}
\alias{image_crop2}
\alias{genPalette}
\title{App Utils}
\usage{
new_waiter(session = NULL)

waiter_html(msg)

is_ci()

is_testing()

get_app_colors()

image_crop2(im, xmin, xmax, ymin, ymax)

genPalette(im)
}
\arguments{
\item{session}{shiny session object}

\item{msg}{message for waiter screen}

\item{im}{image}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}
}
\description{
Functions required to execute and facililate an application user session.
}
\section{Functions}{
\itemize{
\item \code{new_waiter()}: create a new waiter object

\item \code{waiter_html()}: get html for waiter progress page

\item \code{is_ci()}: returns TRUE if called on CI

\item \code{is_testing()}: returns TRUE if called while testing

\item \code{get_app_colors()}: Set Plot colors

\item \code{image_crop2()}: crop image based on plot brush

\item \code{genPalette()}: gen palette plot

}}
